driver: nanohub: chub_ipc: Fix "cast to smaller integer type" errors when compiling
Compiling on Clang 13+ has an additional quality check when compiling. This results to this method error out.

drivers/staging/nanohub/chub_ipc.c:228:7: error: cast to smaller integer type 'u32' (aka 'unsigned int') from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
             if (((u32)ipc_addr[IPC_REG_PERSISTBUF].base + ipc_addr[IPC_REG_PERSISTBUF].offset) >
                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
drivers/staging/nanohub/chub_ipc.c:229:5: error: cast to smaller integer type 'u32' (aka 'unsigned int') from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
                             ((u32)ipc_addr[IPC_REG_IPC].base + ipc_addr[IPC_REG_IPC].offset))

I'm honestly not sure if this is the right cast to do, but it has been tested working and I've done my research and this is the closest we can get on 64-bit arch.

+++ drivers/staging/nanohub/chub_ipc.c
@@ -225,8 +225,8 @@ void *ipc_get_chub_map(void)
 	ipc_addr[IPC_REG_PERSISTBUF].base = ipc_addr[IPC_REG_LOG].base + ipc_addr[IPC_REG_LOG].offset;
 	ipc_addr[IPC_REG_PERSISTBUF].offset = CHUB_PERSISTBUF_SIZE;
 
-	if (((u32)ipc_addr[IPC_REG_PERSISTBUF].base + ipc_addr[IPC_REG_PERSISTBUF].offset) >
-			((u32)ipc_addr[IPC_REG_IPC].base + ipc_addr[IPC_REG_IPC].offset))
+	if (((uintptr_t)ipc_addr[IPC_REG_PERSISTBUF].base + ipc_addr[IPC_REG_PERSISTBUF].offset) >
+			((uintptr_t)ipc_addr[IPC_REG_IPC].base + ipc_addr[IPC_REG_IPC].offset))
 		CSP_PRINTF_INFO("%s: %s: wrong persistbuf addr:%p, %d, ipc_end:0x%x\n",
 			NAME_PREFIX, __func__,
 			ipc_addr[IPC_REG_PERSISTBUF].base, ipc_addr[IPC_REG_PERSISTBUF].offset, map->ipc_end);
